Remove test/dummy from ActiveStorage
The goal here is to remove the committed Rails app that lives in
test/dummy used for initializing Rails and thus ActiveStorage.

Instead, replace it with a test setup only required to load the parts it
is testing.

I re-used the Rails module as a mock for the application for two main
reasons:

1. We need somewhere to house the routes, but that should be fixed in a
   future PR that makes them modular
2. Some migrations depend on Rails config to determine the primary key
   type, there is no public API to get that info since it's applied
   directly to the migration generator during initialization